requirements.txt
flask
gunicorn
google-cloud-vision
google-cloud-storage
google-cloud-datastore
wikipedia







